+++ src/features/study/study-application/ui/StudyApplicationSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
 
-import LoginModal from "@/features/login/components/LoginModal";
+import LoginModal from "@/features/login/ui/LoginModal";
 import Button from "@/shared/components/atoms/Button";
 import useModal from "@/shared/hooks/useModal";
 import ApplyFormCard from "./StudyApplicationFormCard";
@@ -9,10 +9,10 @@ import ApplyFormCard from "./StudyApplicationFormCard";
 import { useAuthStore } from "@/entities/auth/model/auth.store";
 import URL from "@/shared/constants/URL";
 
-import { useParams } from "next/navigation";
 import { useGetQuestions } from "@/entities/study/question/model/question.queries";
 import { isTodayInRange } from "@/features/study/study-application/lib";
 import { StudyApplicationSectionProps } from "@/features/study/study-application/ui/study-application.ui.types";
+import { useParams } from "next/navigation";
 
 const StudyApplicationSection = ({ data }: StudyApplicationSectionProps) => {
   const { id } = useParams() as { id: string };
